+++ Assets/WOM/Scripts/Controller/InsectBullet.cs
@@ -73,7 +73,7 @@ public class InsectBullet : MonoBehaviour
         else
         {
 
-            return GlobalData.instance.statManager.GetInsectMoveSpeed(insectType);
+            return (float)GlobalData.instance.statManager.GetInsectMoveSpeed(insectType);
             // return GlobalData.instance.insectManager.GetInsect(insectType).speed;
         }
     }
